feat(utils): rename service and remove duplicate code
